added support for random integer tensor + tests

+++ src/bnn/random.py
@@ -15,7 +15,7 @@ def generate_random_ternary_tensor(
     dtype: torch.dtype = torch.float,
 ) -> torch.Tensor:
     random = (torch.rand(shape) < desired_var).to(dtype)
-    half = torch.rand_like(random) < 0.5
-    random[half] = -random[half]
+    half = torch.rand_like(random, dtype=float) < 0.5
+    random[half] = -random[half].to(dtype)
 
     return random
